Day 9: add menu Purchase Order and finish all process

+++ database/migrations/2026_05_28_055524_create_sales_order_details_table.php
@@ -32,8 +32,7 @@ return new class extends Migration
             $table->decimal('total_after_ppn', 18, 2)->default(0);
 
             $table->text('keterangan')->nullable();
-            $table->decimal('qty_delivered', 18, 2)
-                ->default(0);
+            $table->integer('qty_delivered')->default(0);
             $table->timestamps();
         });
     }
